Explicit cost, distance, and duration types (#210)
* Explicit cost, distance, and duration types

* Add MeasureType enum class

* Use distance/cost types in prize collecting eval

* Use cost/distance types in operators

* duration_type for time warp, cost_type in evaluator

* Implement dist/dur/cost types almost everywhere on the C++ side

* Add capacity type

* Remove direct use of INT_MAX macros

* SWAP* deltaDist

* operator/=

* ostream operator<< for Measures

* Explicit type for std::max

* Rename distance_type to Distance

* Rename duration_type to Duration

* Rename cost_type to Cost

* Rename capacity_type to Capacity

* Rename value_type to Value

* Rename Capacity to Load, update shortcut condition in LS::maybeInsert

* Hash measures

* Implement Coordinate

* Move type aliases up, before implementation details

* Specialise std::numeric_limits for strong types

* Implement Measure::get()

+++ pyvrp/cpp/crossover/selective_route_exchange.cpp
@@ -20,8 +20,8 @@ double routeAngle(ProblemData const &data, Route const &route)
     // angle, but is much faster to compute. See the following post for details:
     // https://stackoverflow.com/a/16561333/4316405.
     auto const [routeX, routeY] = route.centroid();
-    auto const dx = routeX - data.depot().x;
-    auto const dy = routeY - data.depot().y;
+    auto const dx = routeX - static_cast<double>(data.depot().x);
+    auto const dy = routeY - static_cast<double>(data.depot().y);
     return std::copysign(1. - dx / (std::fabs(dx) + std::fabs(dy)), dy);
 }
